Popravi grešku s date() funkcijom u predmet_helper.class.php

+++ seup/class/predmet_helper.class.php
@@ -443,9 +443,9 @@ class Predmet_helper
                 // Date formatting
                 $date_formatted = 'N/A';
                 if (isset($doc->date_c)) {
-                    $date_formatted = date('d.m.Y H:i', $doc->date_c);
+                    $date_formatted = date('d.m.Y H:i', (int)$doc->date_c);
                 } elseif (isset($doc->last_modified)) {
-                    $date_formatted = date('d.m.Y H:i', strtotime($doc->last_modified));
+                    $date_formatted = date('d.m.Y H:i', (int)strtotime($doc->last_modified));
                 }
                 $documentTableHTML .= '<td><div class="seup-document-date"><i class="fas fa-calendar me-1"></i>' . $date_formatted . '</div></td>';
